requirements.txt
pbr>=0.5.21,<1.0

eventlet
gearman>=2.0.2
oslo.config>=1.2.0
python-daemon>=1.6
python-logstash
python_novaclient>=2.14.1,<2.14.2
python_swiftclient>=1.3.0
requests>=1.0.0
dogapi
pecan
sqlalchemy>=0.8.0
wsme>=0.5b2
mysql-connector-python
ipaddress==1.0.4
six<1.4.0
futures
babel
